safe_malloc, safe_malloc argument type become int32_t

+++ sperl_allocator_util.h
@@ -1,7 +1,7 @@
 #ifndef SPERL_ALLOCATOR_UTIL_H
 #define SPERL_ALLOCATOR_UTIL_H
 
-void* SPerl_ALLOCATOR_UTIL_safe_malloc(size_t count, size_t size);
-void* SPerl_ALLOCATOR_UTIL_safe_realloc(void* ptr, size_t count, size_t size);
+void* SPerl_ALLOCATOR_UTIL_safe_malloc(int32_t count, int32_t size);
+void* SPerl_ALLOCATOR_UTIL_safe_realloc(void* ptr, int32_t count, int32_t size);
 
 #endif
